% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{assets}
\alias{assets}
\alias{format.assets}
\alias{print.assets}
\alias{[.assets}
\alias{as.double.assets}
\alias{names.assets}
\alias{as.data.frame.assets}
\title{Assets}
\usage{
assets(x)

\method{format}{assets}(x, ...)

\method{print}{assets}(x, ...)

\method{[}{assets}(x, i, ...)

\method{as.double}{assets}(x, ...)

\method{names}{assets}(x)

\method{as.data.frame}{assets}(x, row.names = NULL, optional = NA, ...)
}
\arguments{
\item{x}{object of class 'assets'.}

\item{\dots}{currently ignored.}

\item{i}{elements to extract or replace. Numeric values coerced
to integer identifying the \code{i}th element(s) for subsetting.}

\item{row.names}{\code{NULL} (default) or a character vector giving the row names for
the data frame. Only used if \code{x} is of length 1L.}

\item{optional}{currently ignored.}
}
\value{
An asset data frame with the name of the asset and all provided
additional information.
}
\description{
Each item can provide multiple assets. An asset specifies the type of data
(e.g., historical, recent) as well as the temporal resolution of the data
and the data period (if suitable). The \code{ms_assets} class is used to handle
these assets within the package.
}
\author{
Reto
}
