% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{ms_collections}
\alias{ms_collections}
\title{Available Collections}
\usage{
ms_collections(verbose = FALSE, raw = FALSE)
}
\arguments{
\item{verbose}{logical, defaults to \code{FALSE}. If set \code{TRUE} the
some messages are printed.}

\item{raw}{logical, defaults to \code{FALSE} (see Return).}
}
\value{
A data frame with a series of properties from the
collections (if \code{raw = FALSE}), else a list of lists with
the unformatted return from the API. Each element of the
list is the result of one API request (paging).
}
\description{
Requesting a list of all available collections provided via the API.
}
\examples{
\dontrun{
## Fetch all available collections
collections <- ms_collections()

## Extract collections by MeteoSchweiz
subset(res, grepl("meteoschweiz", id))
}

}
\author{
Reto
}
