% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{sg_api_url}
\alias{sg_api_url}
\title{Generate Geoportal API URLs}
\usage{
sg_api_url(...)
}
\arguments{
\item{...}{additional arguments to extend the base API URL.
All arguments (after coercion to character) must represent
valid non-empty characters.}
}
\value{
Character of length one with the URL of the API end point.
}
\description{
Generate Geoportal API URLs
}
\details{
On loading the package sets the option \code{"swissgeo.apiurl"}
with the main entry point to the API. This is used inside this
function (i.e., by the entire package) to talk to the API.

This would allow users to change the end-point (or version)
without updating the package, e.g., for testing.
}
\examples{
## Base URL
sg_api_url()

## Extending the base URL to point to a specific API endpoint
sg_api_url("ch.meteoschweiz.ogd-smn")
sg_api_url("ch.meteoschweiz.ogd-smn", "items")

## Changing default API URL. Possible, but not a standard use-case.
hold_apiurl <- getOption("swissgeo.apiurl") # kept for resetting
options(swissgeo.apiurl = "https://some.new.domain/api/stac/v5")
sg_api_url()

## Setting back to defaults
options(swissgeo.apiurl = hold_apiurl)
sg_api_url()

}
\author{
Reto
}
