% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessapi.R
\name{sg_assets}
\alias{sg_assets}
\title{Collection Assets}
\usage{
sg_assets(id, verbose = FALSE, raw = FALSE)
}
\arguments{
\item{id}{character, ID of the collection for which to retrieve
the items (see \code{\link[=sg_collections]{sg_collections()}}).}

\item{verbose}{logical, defaults to \code{FALSE}. If set \code{TRUE} the
some messages are printed.}

\item{raw}{logical, defaults to \code{FALSE} (see Return).}
}
\value{
A tbl data frame containing the asset details is returned.

If \code{raw = TRUE} a list of lists is returned with the raw (json decoded)
result from the API calls. Each element in the list corresponds to one API
call (paging).
}
\description{
Retrieving all assets of a specific collection.
}
\author{
Reto
}
